+++ docs/source/conf.py
@@ -17,7 +17,7 @@ Notes:
     # need to edit the conf.py
 
     cd ~/code/ndsampler/docs
-    sphinx-apidoc --private --separate -f -o ~/code/ndsampler/docs/source/auto ~/code/ndsampler/ndsampler
+    sphinx-apidoc --private --separate --force --output-dir ~/code/ndsampler/docs/source/auto ~/code/ndsampler/ndsampler
 
     # Note: the module should importable before running this
     # (e.g. install it in developer mode or munge the PYTHONPATH)
@@ -136,7 +136,7 @@ def parse_version(fpath):
     return visitor.version
 
 project = 'ndsampler'
-copyright = '2024, Jon Crall'
+copyright = '2025, Jon Crall'
 author = 'Jon Crall'
 modname = 'ndsampler'
 
@@ -217,7 +217,7 @@ intersphinx_mapping = {
     # 'xxhash': ('https://pypi.org/project/xxhash/', None),
     # 'pygments': ('https://pygments.org/docs/', None),
     # 'tqdm': ('https://tqdm.github.io/', None),
-    # Requries that the repo have objects.inv
+    # Requires that the repo have objects.inv
     'kwarray': ('https://kwarray.readthedocs.io/en/latest/', None),
     'kwimage': ('https://kwimage.readthedocs.io/en/latest/', None),
     # 'kwplot': ('https://kwplot.readthedocs.io/en/latest/', None),
@@ -588,7 +588,7 @@ class GoogleStyleDocstringProcessor:
 
             accum.append(line)
 
-        # Finialize the last section
+        # Finalize the last section
         accept()
 
         lines[:] = new_lines